+++ test/Serilog.Sinks.HttpTests.LogServer/Program.cs
@@ -1,0 +1,11 @@
+using Serilog.Sinks.HttpTests.LogServer;
+
+var builder = Host
+    .CreateDefaultBuilder(args)
+    .ConfigureWebHostDefaults(webBuilder =>
+    {
+        webBuilder.UseStartup<Startup>();
+    });
+
+using var app = builder.Build();
+app.Run();