+++ test/Serilog.Sinks.HttpTests.LogServer/Startup.cs
@@ -1,0 +1,35 @@
+﻿using Serilog.Sinks.HttpTests.LogServer.Services;
+
+namespace Serilog.Sinks.HttpTests.LogServer;
+
+public class Startup
+{
+    public Startup(IConfiguration configuration)
+    {
+        Configuration = configuration;
+    }
+
+    public IConfiguration Configuration { get; }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddControllers();
+        services.AddSingleton<HealthService>();
+        services.AddSingleton<LogEventService>();
+    }
+
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
+        {
+            app.UseDeveloperExceptionPage();
+        }
+
+        app.UseRouting();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllers();
+        });
+    }
+}